requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.30.0
ollama>=0.1.0
Pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
albumentations>=1.3.0
opencv-python>=4.7.0
deep-translator>=1.11.0
gTTS>=2.3.0
